Change timeouts to INT, refactoring

+++ adapters/saml/core/src/main/java/org/keycloak/adapters/cloned/AdapterHttpClientConfig.java
@@ -76,11 +76,11 @@ public interface AdapterHttpClientConfig {
     /**
      * Returns timeout for socket waiting for data in milliseconds.
      */
-    long getSocketTimeout();
+    int getSocketTimeout();
 
     /**
      * Returns timeout for establishing the connection with the remote host in milliseconds.
      */
-    long getConnectionTimeout();
+    int getConnectionTimeout();
 
 }
